Abstract OpenAI into protocol; make test depend on mocks
