+++ Sources/OpenAI/Public/Extensions/OpenAI+Extensions.swift
@@ -75,5 +75,4 @@ public extension OpenAI {
             }
         }
     }
-
 }
