+++ Sources/OpenAI/Public/Extensions/URLSession+Extensions.swift
@@ -1,0 +1,16 @@
+//
+//  URLSession+Extensions.swift
+//  
+//
+//  Created by Sihao Lu on 3/30/23.
+//
+
+import Foundation
+
+extension URLSession: URLSessionProtocol {
+    public func dataTask(with request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTaskProtocol {
+        return dataTask(with: request, completionHandler: completionHandler) as URLSessionDataTask
+    }
+}
+
+extension URLSessionDataTask: URLSessionDataTaskProtocol {}
