+++ Tests/OpenAITests/OpenAITests.swift
@@ -1,34 +1,75 @@
+//
+//  OpenAIImplTests.swift
+//  
+//
+//  Created by Sihao Lu on 3/30/23.
+//
+
 import XCTest
 @testable import OpenAI
 
-@available(iOS 13.0, *)
-@available(watchOS 6.0, *)
-@available(tvOS 13.0, *)
-final class OpenAITests: XCTestCase {
+final class OpenAIImplTests: XCTestCase {
 
     var openAI: OpenAI!
+    var mockURLSession: MockURLSession!
 
     override func setUp() {
         super.setUp()
-        self.openAI = OpenAI(apiToken: "<YOUR TOKEN HERE>")
+        self.mockURLSession = MockURLSession()
+        self.openAI = OpenAIImpl(apiToken: "<MOCK_API_TOKEN>", session: mockURLSession)
     }
 
     func testCompletionsAsync() async throws {
         let query = OpenAI.CompletionsQuery(model: .textDavinci_003, prompt: "What is 42?", temperature: 0, maxTokens: 100, topP: 1, frequencyPenalty: 0, presencePenalty: 0, stop: ["\\n"])
+        mockURLSession.mockData = "{\"id\":\"cmpl-6zyoigxENXmLGDbEE43Un50UmKTFs\",\"object\":\"text_completion\",\"created\":1680229808,\"model\":\"text-davinci-003\",\"choices\":[{\"text\":\"\\n\\n42 is the answer to the ultimate question of life, the universe, and everything, according to the book The Hitchhiker\'s Guide to the Galaxy.\",\"index\":0,\"logprobs\":null,\"finish_reason\":\"stop\"}],\"usage\":{\"prompt_tokens\":4,\"completion_tokens\":33,\"total_tokens\":37}}\n".data(using: .utf8)
         let result = try await openAI.completions(query: query)
+
+        XCTAssertEqual(mockURLSession.lastRequest?.allHTTPHeaderFields, ["Content-Type": "application/json", "Authorization": "Bearer <MOCK_API_TOKEN>"])
+        let httpBody = try XCTUnwrap(mockURLSession.lastRequest?.httpBody.map { String(data: $0, encoding: .utf8) })
+        XCTAssertEqual(
+            httpBody,
+            "{\"prompt\":\"What is 42?\",\"frequency_penalty\":0,\"model\":\"text-davinci-003\",\"stop\":[\"\\\\n\"],\"temperature\":0,\"presence_penalty\":0,\"top_p\":1,\"max_tokens\":100}"
+        )
         XCTAssertFalse(result.choices.isEmpty)
     }
-    
+
     func testImages() async throws {
         let query = OpenAI.ImagesQuery(prompt: "White cat with heterochromia sitting on the kitchen table", n: 1, size: "1024x1024")
+
+        mockURLSession.mockData = "{\n  \"created\": 1680230109,\n  \"data\": [\n    {\n      \"url\": \"https://oaidalleapiprodscus.blob.core.windows.net/private/org-rRP9dUDCiFLX0N45fYzo6P6K/user-IUShDLHEZxJnCo8QMrkdzcCN/img-8rOmgcr3sP1dWuyHlVqsfhhY.png?st=2023-03-31T01%3A35%3A09Z&se=2023-03-31T03%3A35%3A09Z&sp=r&sv=2021-08-06&sr=b&rscd=inline&rsct=image/png&skoid=6aaadede-4fb3-4698-a8f6-684d7786b067&sktid=a48cca56-e6da-484e-a814-9c849652bcb3&skt=2023-03-30T15%3A57%3A30Z&ske=2023-03-31T15%3A57%3A30Z&sks=b&skv=2021-08-06&sig=CNED0Fl%2BaMjeBHzgUHVMBxci9DRUkpe6LcsC1t9AJDk%3D\"\n    }\n  ]\n}\n".data(using: .utf8)
+
         let result = try await openAI.images(query: query)
+        XCTAssertEqual(mockURLSession.lastRequest?.allHTTPHeaderFields, ["Content-Type": "application/json", "Authorization": "Bearer <MOCK_API_TOKEN>"])
+        let httpBody = try XCTUnwrap(mockURLSession.lastRequest?.httpBody.map { String(data: $0, encoding: .utf8) })
+        XCTAssertEqual(
+            httpBody,
+            "{\"n\":1,\"size\":\"1024x1024\",\"prompt\":\"White cat with heterochromia sitting on the kitchen table\"}"
+        )
         XCTAssertFalse(result.data.isEmpty)
     }
 
     func testEmbeddings() async throws {
         let query = OpenAI.EmbeddingsQuery(model: .textSearchBabbadgeDoc, input: "The food was delicious and the waiter...")
+
+        mockURLSession.mockData = "{\n  \"object\": \"list\",\n  \"data\": [\n    {\n      \"object\": \"embedding\",\n      \"index\": 0,\n      \"embedding\": [\n        0.0010535449,\n        0.024234328,\n        -0.0084999,\n        0.008647452,\n        0.027292684,\n        -0.007471506,\n        0.00040884252,\n        0.026809786,\n        0.015533221,\n        -0.029009653,\n        -0.00042421254,\n        0.006841056,\n        0.011696864,\n        0.0259513,\n        0.021980803,\n        -0.009926238,\n        0.0073507815,\n        -0.018385896,\n        0.020031327,\n        0.04761017,\n        0.07597596,\n        0.029599862,\n        0.0007427911,\n        -0.009926238,\n        -0.017214421,\n        -0.00608541,\n        -0.045106255,\n        -0.017625779,\n        -0.051366046,\n        -0.024359524,\n        -0.0008556909,\n        0.02452049,\n        -0.0023943696,\n        -0.0147105055,\n        0.031567223,\n        -0.0057277074,\n        0.032908607,\n        -0.021462137,\n        -0.0029331585,\n        -0.02108655,\n        -0.0074446783,\n        -0.0028951527,\n        0.030136416,\n        -0.018457437,\n        -0.029617747,\n        0.030458348,\n        0.0044466862,\n        0.039991114,\n        -0.036217354,\n        0.01535437,\n        -0.017393272,\n        0.026917096,\n        0.014826759,\n        0.0011781818,\n        0.045356646,\n        0.026917096,\n        0.00043818526,\n        0.036396205,\n        0.023876626,\n        0.037522968,\n        0.016740466,\n        -0.03278341,\n        0.037522968,\n        0.016346993,\n        0.0023451855,\n        0.041135762,\n        0.01885091,\n        0.015864095,\n        0.003085182,\n        0.013145559,\n        0.008137726,\n        -0.008964913,\n        0.034679234,\n        0.03546618,\n        0.040098425,\n        -0.00079476973,\n        0.0027766638,\n        -0.022749864,\n        0.01745587,\n        -0.0010859617,\n        0.025074929,\n        -0.0301543,\n        -0.008307635,\n        0.03575234,\n        0.0049988893,\n        -0.03011853,\n        -0.019673625,\n        0.022642553,\n        -0.022338506,\n        0.03278341,\n        0.010900976,\n        -0.017482698,\n        0.017679434,\n        -0.0063849855,\n        0.009496995,\n        0.005468373,\n        0.046644375,\n        0.01675835,\n        0.014549539,\n        0.0004077247,\n        0.005092786,\n        0.0044824565,\n        0.0267919,\n        -0.00011031482,\n        -0.03340939,\n        -0.043925837,\n        0.006872355,\n        0.049720615,\n        -0.0019986613,\n        0.014442229,\n        0.005155384,\n        -0.025343206,\n        -0.024180673,\n        -0.0034294706,\n        -0.025718793,\n        0.015551106,\n        0.039061088,\n        -0.011938313,\n        -0.022266965,\n        0.013127673,\n        -0.013243927,\n        0.00025025185,\n        -0.035448294,\n        -0.025450516,\n        -0.013449606,\n        0.015604761,\n        -0.020371145,\n        0.01507715,\n        -0.008388118,\n        0.022517357,\n        -0.026774015,\n        0.023697775,\n        0.023465268,\n        -0.021623101,\n        -0.03798798,\n        0.032407824,\n        0.008575912,\n        0.01266266,\n        0.007967818,\n        -0.017590009,\n        0.027042292,\n        -0.018099735,\n        -0.046751685,\n        -0.019047646,\n        -0.0027766638,\n        -0.029993335,\n        0.010883091,\n        0.028276363,\n        0.003599379,\n        -0.005660638,\n        0.038524535,\n        -0.018135505,\n        -0.007149574,\n        0.012304958,\n        0.012117164,\n        0.017357502,\n        0.0028101983,\n        -0.020657307,\n        0.017008742,\n        0.005606983,\n        0.014612137,\n        -0.047860563,\n        0.006912596,\n        0.0005583509,\n        0.014674735,\n        0.016177084,\n        0.0018790547,\n        0.03424999,\n        0.013977216,\n        -0.02305391,\n        0.025861874,\n        -0.001498996,\n        0.0030136416,\n        -0.024985503,\n        0.025718793,\n        0.037272576,\n        0.02840156,\n        0.019834591,\n        0.013056133,\n        -0.011312334,\n        -0.00982787,\n        -0.0408496,\n        0.011115598,\n        -0.01961997,\n        -0.03780913,\n        -0.010516447,\n        0.021694642,\n        0.0013246163,\n        -0.006139065,\n        -0.019280152,\n        0.03187127,\n        -0.048182495,\n        0.007815794,\n        0.0054907296,\n        -0.00118936,\n        0.01672258,\n        -0.04467701,\n        -0.03938302,\n        -0.014960897,\n        0.01665104,\n        0.0008909022,\n        0.016042946,\n        -0.015112921,\n        0.0030337623,\n        -0.016990857,\n        -0.05193837,\n        -0.01332441,\n        0.018582633,\n        -0.021068664,\n        -0.0031947284,\n        0.0068947114,\n        -0.029868139,\n        0.0018969398,\n        0.026541509,\n        -0.022159655,\n        0.052832626,\n        -0.022553127,\n        0.003822943,\n        0.012412269,\n        -0.0060630534,\n        0.030512003,\n        -0.0144064585,\n        0.013682112,\n        -0.0068142284,\n        -0.021140205,\n        0.0017996895,\n        -0.0063671004,\n        -0.011678979,\n        -0.015273887,\n        0.019870361,\n        -0.01539014,\n        0.025128584,\n        -0.007538575,\n        0.000724906,\n        -0.003617264,\n        -0.0028817388,\n        -0.0034339419,\n        -0.019709395,\n        -0.004310312,\n        -0.00416276,\n        0.0040666275,\n        -0.045642808,\n        0.019208612,\n        -0.015631588,\n        -0.036843333,\n        -0.013583744,\n        0.028777147,\n        -0.017750975,\n        0.022767749,\n        -0.0011217319,\n        0.0015370019,\n        0.010900976,\n        -0.0021763947,\n        -0.021730412,\n        -0.018833024,\n        0.030279497,\n        -0.016275452,\n        0.020263834,\n        0.0013458547,\n        0.010784723,\n        -0.01822493,\n        0.021462137,\n        -0.014782046,\n        0.013351237,\n        -0.0129845925,\n        0.026469968,\n        -0.002336243,\n        0.04439085,\n        0.033749208,\n        -0.020621536,\n        0.00035071588,\n        0.004473514,\n        -0.027614616,\n        -0.0105879875,\n        0.05991513,\n        -0.0062553184,\n        0.010659528,\n        0.004945234,\n        -0.01423655,\n        -0.032175317,\n        -0.022106,\n        -0.0047440263,\n        0.012367556,\n        -0.029438896,\n        0.01539014,\n        0.0012240125,\n        0.045070484,\n        0.03759451,\n        0.007967818,\n        -0.01633805,\n        0.035770226,\n        -0.0025195654,\n        -0.023179106,\n        0.025468402,\n        0.022195425,\n        -0.03945456,\n        0.023930281,\n        -0.0055667413,\n        0.03260456,\n        -0.00986364,\n        0.02777558,\n        -0.0060988236,\n        0.0071719303,\n        -0.0022624668,\n        0.007932047,\n        0.015801497,\n        0.07375821,\n        -0.018493207,\n        -0.02598707,\n        0.0025687495,\n        -0.0046143592,\n        0.0023250647,\n        0.0032930966,\n        0.01374471,\n        -0.03543041,\n        0.008057243,\n        -0.014862529,\n        -0.05369111,\n        -0.00577242,\n        0.0018611696,\n        -0.019924017,\n        0.034965396,\n        0.003921311,\n        -0.02459203,\n        0.0040666275,\n        0.009577478,\n        -0.014191837,\n        0.04417623,\n        0.0022144006,\n        -0.018582633,\n        0.005124085,\n        -0.0004348318,\n        0.0074938624,\n        0.010623758,\n        0.0034249993,\n        -0.012135049,\n        0.013172386,\n        -0.013539031,\n        -0.018940335,\n        0.0071451026,\n        -0.0025508644,\n        -0.016311223,\n        -0.015309657,\n        0.018546863,\n        0.030798165,\n        0.020621536,\n        -0.006232962,\n        0.012725258,\n        0.012367556,\n        0.03539464,\n        0.016060831,\n        0.00026296705,\n        -0.009157178,\n        0.02217754,\n        0.0047619115,\n        0.022606783,\n        -0.031245293,\n        0.05537231,\n        0.016704695,\n        0.010042491,\n        0.009729502,\n        0.0041292254,\n        -0.014173952,\n        -0.03480443,\n        -0.0023273004,\n        -0.016901432,\n        0.0008484251,\n        -0.0045875316,\n        0.0075296327,\n        -0.019387463,\n        -0.00986364,\n        -0.027936546,\n        0.0063849855,\n        0.000144059,\n        -0.029152734,\n        0.008137726,\n        -0.01787617,\n        -0.03909686,\n        0.017393272,\n        0.011491185,\n        -0.020263834,\n        0.00028169053,\n        0.023232762,\n        -0.024627801,\n        -0.020907698,\n        0.01266266,\n        0.062740974,\n        0.0014788753,\n        -0.024198558,\n        -0.014075584,\n        -0.0091169365,\n        -0.012913052,\n        -0.005848432,\n        -0.006617492,\n        -0.0047306125,\n        0.023214877,\n        0.00853567,\n        -0.009488053,\n        0.0063536866,\n        -0.04431931,\n        0.025486287,\n        -0.026004955,\n        -0.037129495,\n        0.004846866,\n        -0.016355935,\n        0.023912396,\n        0.0195842,\n        -0.0019282388,\n        0.07297126,\n        0.009649019,\n        -0.008374704,\n        0.021068664,\n        0.026523624,\n        0.041958477,\n        0.0033914647,\n        -0.008566969,\n        0.01902976,\n        0.013807308,\n        -0.02675613,\n        -0.012564292,\n        -0.007471506,\n        -0.019941902,\n        0.028079627,\n        -0.00053627393,\n        0.02305391,\n        -0.005624868,\n        -0.00489605,\n        0.0077844956,\n        -0.0055712126,\n        0.034536153,\n        -0.0013838606,\n        -0.029671403,\n        -0.031173753,\n        0.007999117,\n        -0.0032439125,\n        0.0195842,\n        -0.024091247,\n        -0.026309002,\n        0.013226042,\n        0.0013458547,\n        -0.023518924,\n        -0.019137071,\n        0.013163444,\n        0.030547773,\n        0.013199214,\n        0.003543488,\n        0.03340939,\n        0.040384587,\n        -0.02046057,\n        0.02221331,\n        0.03197858,\n        -0.028991768,\n        -0.0014632258,\n        0.006250847,\n        -0.019637855,\n        -0.013789423,\n        0.01577467,\n        0.021748297,\n        -0.012510637,\n        0.011312334,\n        -0.0055533275,\n        -0.001028394,\n        -0.043460824,\n        0.0014252199,\n        0.032300513,\n        -0.01255535,\n        -0.017366445,\n        -0.009237661,\n        -0.02308968,\n        -0.01745587,\n        0.02997545,\n        0.015944578,\n        0.048397116,\n        -0.035966963,\n        -0.003995087,\n        0.01976305,\n        0.011875715,\n        -0.010820493,\n        0.010400194,\n        0.0152381165,\n        0.011839945,\n        -0.027543075,\n        0.010892034,\n        -0.0064475834,\n        0.00043343453,\n        -0.0024390824,\n        0.021748297,\n        -0.01630228,\n        0.003203671,\n        -0.013351237,\n        -0.004435508,\n        0.038560305,\n        -0.014844644,\n        0.007471506,\n        0.009693732,\n        -0.056051943,\n        -0.0041269897,\n        0.0097921,\n        -0.031334717,\n        0.023125451,\n        -0.00022090909,\n        0.041994248,\n        -0.036879104,\n        0.007422322,\n        -0.02283929,\n        0.016132372,\n        0.013208156,\n        0.011464357,\n        -0.009201891,\n        0.020245949,\n        -0.014495884,\n        0.016472189,\n        0.0061435364,\n        0.011294449,\n        -0.018421667,\n        0.03773759,\n        0.0007662653,\n        0.012000911,\n        0.006326859,\n        0.032246858,\n        -0.00004768199,\n        0.01293988,\n        0.0007137278,\n        -0.0028906814,\n        0.035895422,\n        0.012358613,\n        0.019262267,\n        0.005235867,\n        -0.026523624,\n        -0.00713616,\n        0.008428359,\n        0.016087659,\n        0.0013112024,\n        0.01640959,\n        -0.03169242,\n        0.0339996,\n        -0.026487853,\n        0.004299134,\n        -0.03655717,\n        -0.01017663,\n        0.0040152078,\n        -0.006188249,\n        0.00913035,\n        0.045535497,\n        -0.023447383,\n        -0.007985703,\n        -0.11110232,\n        0.00086631015,\n        0.008236094,\n        0.06152479,\n        0.033910174,\n        -0.037308346,\n        -0.008701107,\n        0.007359724,\n        -0.004095691,\n        0.011607438,\n        -0.0111781955,\n        0.0045249336,\n        0.012993535,\n        -0.0036664482,\n        -0.020871928,\n        -0.022374276,\n        0.0025955772,\n        0.040134195,\n        0.006071996,\n        -0.009246604,\n        0.0042790133,\n        0.020818273,\n        -0.014826759,\n        -0.0028325547,\n        0.025933415,\n        -0.03074451,\n        -0.01626651,\n        -0.021319056,\n        0.01881514,\n        0.0139593305,\n        0.0036686838,\n        -0.005012303,\n        0.016892489,\n        0.03575234,\n        0.0097563295,\n        0.021784067,\n        0.012117164,\n        -0.0053968327,\n        -0.00407557,\n        -0.05029294,\n        0.021480022,\n        -0.08126996,\n        0.024985503,\n        0.00018639641,\n        0.026738245,\n        -0.036986414,\n        -0.012358613,\n        0.015515336,\n        -0.047002077,\n        0.068964995,\n        -0.020800387,\n        0.00013909867,\n        0.0071719303,\n        -0.01976305,\n        0.017303847,\n        0.021891378,\n        0.021032894,\n        0.009926238,\n        0.027364224,\n        0.00011834915,\n        0.0017203244,\n        0.0040174434,\n        -0.0043460825,\n        -0.015944578,\n        -0.021354826,\n        -0.015408025,\n        0.022481587,\n        -0.0125374645,\n        -0.041851167,\n        -0.00986364,\n        0.03308746,\n        -0.026362658,\n        0.0042275935,\n        -0.05476422,\n        -0.0005779127,\n        -0.001121173,\n        0.017339617,\n        0.004752969,\n        0.02291083,\n        -0.02741788,\n        0.001951713,\n        -0.007923105,\n        -0.013762595,\n        -0.0039548455,\n        0.017178651,\n        0.019208612,\n        -0.007744254,\n        0.029760828,\n        -0.035770226,\n        0.013530089,\n        0.010015664,\n        -0.0007522926,\n        -0.009273431,\n        -0.035180017,\n        -0.02661305,\n        -0.006747159,\n        -0.009175063,\n        -0.0027453648,\n        -0.018403782,\n        0.047002077,\n        0.0060004555,\n        0.00007321718,\n        -0.034017485,\n        0.005235867,\n        -0.02105078,\n        0.04453393,\n        -0.016203912,\n        0.008911258,\n        -0.028133282,\n        0.023357958,\n        0.017929826,\n        -0.01602506,\n        0.025504172,\n        -0.013145559,\n        0.030172186,\n        -0.004536112,\n        0.015059265,\n        0.016606327,\n        0.010704241,\n        0.014093469,\n        0.048504427,\n        -0.0074133794,\n        0.0013033777,\n        0.008906786,\n        -0.0034361775,\n        0.020013442,\n        0.010319711,\n        -0.031191638,\n        -0.0056159254,\n        0.0011228498,\n        0.019476889,\n        -0.019494774,\n        -0.022266965,\n        -0.00084171817,\n        0.065709904,\n        0.020192293,\n        0.0036932759,\n        -0.005150913,\n        0.04306735,\n        -0.012430154,\n        -0.019494774,\n        -0.0022524064,\n        0.0019293566,\n        -0.008812889,\n        -0.042030018,\n        -0.018403782,\n        0.0016007177,\n        0.0079499325,\n        0.023214877,\n        0.0009881525,\n        0.01056116,\n        -0.0037961153,\n        0.0105879875,\n        0.025164355,\n        -0.0056203967,\n        -0.04077806,\n        0.004976533,\n        -0.021748297,\n        -0.0051017287,\n        0.030494118,\n        -0.032658216,\n        0.005168798,\n        -0.030297382,\n        0.01811762,\n        -0.031406257,\n        0.010650585,\n        0.0003426117,\n        0.003098596,\n        -0.00022873383,\n        -0.004869222,\n        0.015309657,\n        -0.002774428,\n        0.008146669,\n        0.013333352,\n        -0.03641409,\n        -0.021175975,\n        0.0038676558,\n        0.0127610285,\n        0.0041091046,\n        -0.027632501,\n        -0.0017460341,\n        -0.01077578,\n        0.033838633,\n        0.0008931378,\n        -0.002479324,\n        -0.038274143,\n        0.046823226,\n        0.015086093,\n        0.047002077,\n        -0.0031254236,\n        -0.0049899467,\n        0.024270099,\n        -0.0055041434,\n        -0.0025262723,\n        0.02584399,\n        0.0147909885,\n        -0.027435765,\n        -0.024323754,\n        0.0532261,\n        -0.03959764,\n        0.008343405,\n        0.024913963,\n        0.015676301,\n        0.025432631,\n        -0.0046635433,\n        0.006791872,\n        -0.026899211,\n        -0.0041023977,\n        0.011419645,\n        -0.0071942867,\n        0.02781135,\n        0.024913963,\n        0.011106655,\n        -0.0005105641,\n        -0.024019707,\n        -0.011401759,\n        0.009979893,\n        -0.0021227393,\n        0.018833024,\n        0.0041158115,\n        0.00853567,\n        -0.015193404,\n        0.009461225,\n        0.009063281,\n        0.01976305,\n        0.030512003,\n        0.026040725,\n        0.009175063,\n        0.013699997,\n        0.062490586,\n        -0.00031494568,\n        -0.03179973,\n        0.0073373676,\n        -0.03938302,\n        -0.0082539795,\n        0.03627101,\n        -0.041350383,\n        0.018797254,\n        0.0027207728,\n        0.026988637,\n        0.00035295152,\n        0.01430809,\n        0.014039814,\n        0.003970495,\n        -0.0028884457,\n        -0.02759673,\n        -0.23966052,\n        0.04886213,\n        0.005285051,\n        -0.052331842,\n        0.011240793,\n        -0.007019907,\n        0.021927148,\n        -0.0076995413,\n        -0.0106327,\n        -0.011097712,\n        -0.014638965,\n        0.018421667,\n        0.0065459516,\n        0.010847321,\n        -0.014719448,\n        -0.00058406073,\n        -0.01633805,\n        0.016749408,\n        0.0049362914,\n        0.0046546008,\n        -0.018368011,\n        -0.0029219803,\n        -0.0043527894,\n        0.017831458,\n        -0.0034831257,\n        -0.010078262,\n        -0.0038698914,\n        0.0033534588,\n        -0.0165795,\n        0.01056116,\n        0.041600775,\n        -0.011500128,\n        0.00815114,\n        0.03264033,\n        0.034375187,\n        -0.027113833,\n        -0.013547974,\n        -0.005942329,\n        0.010087204,\n        0.021140205,\n        0.00070031395,\n        -0.024806652,\n        -0.029492551,\n        -0.011518013,\n        -0.0047171987,\n        0.019476889,\n        -0.017616836,\n        0.0085580265,\n        -0.011455415,\n        0.02602284,\n        0.017849343,\n        -0.018368011,\n        0.01024817,\n        -0.022588897,\n        0.0019964257,\n        -0.008714521,\n        0.034196336,\n        0.024502605,\n        -0.010918861,\n        -0.0017091462,\n        -0.002105972,\n        -0.013100846,\n        0.0032304986,\n        -0.021426367,\n        -0.005660638,\n        -0.0014196308,\n        0.015184461,\n        -0.00026059168,\n        0.01297565,\n        0.0070780334,\n        -0.022553127,\n        0.0022009867,\n        0.016847776,\n        -0.019995557,\n        0.009640076,\n        -0.0035412523,\n        0.0033042748,\n        -0.013530089,\n        0.015264944,\n        0.011831002,\n        0.0010043608,\n        -0.018368011,\n        0.004006265,\n        0.03385652,\n        0.015622647,\n        -0.021408482,\n        0.029850254,\n        0.036664482,\n        0.014281263,\n        -0.012895167,\n        -0.0020903226,\n        -0.01766155,\n        0.010731068,\n        -0.017634721,\n        0.012859397,\n        -0.013100846,\n        -0.012188705,\n        -0.015229174,\n        0.0020634949,\n        0.022892945,\n        -0.012054566,\n        0.011383874,\n        -0.014638965,\n        -0.03428576,\n        -0.0063045025,\n        -0.040098425,\n        -0.022445817,\n        -0.020728847,\n        0.0030270554,\n        0.00020525962,\n        0.0076235295,\n        0.036467746,\n        -0.0075788167,\n        0.025593597,\n        -0.027310569,\n        -0.051366046,\n        0.009908353,\n        0.020192293,\n        0.011205023,\n        0.03644986,\n        -0.03256879,\n        -0.015935635,\n        0.010766838,\n        -0.011115598,\n        -0.032032236,\n        0.03396383,\n        0.001016098,\n        -0.018940335,\n        0.035197902,\n        -0.040348817,\n        0.008777119,\n        -0.034124795,\n        -0.0003370226,\n        -0.013038248,\n        -0.012161877,\n        -0.00021266518,\n        0.0008031534,\n        -0.024931848,\n        0.022606783,\n        -0.024466835,\n        -0.00940757,\n        0.0051419702,\n        -0.0073015974,\n        0.014021928,\n        -0.012179762,\n        0.001362622,\n        0.023572579,\n        -0.00028434535,\n        0.012385441,\n        -0.01150907,\n        -0.016883546,\n        0.022088114,\n        -0.0031813146,\n        0.004399738,\n        0.02598707,\n        -0.0049362914,\n        0.010400194,\n        0.016445361,\n        0.017831458,\n        0.0019863653,\n        -0.015631588,\n        0.0008702225,\n        -0.016642097,\n        -0.011374932,\n        -0.018251758,\n        0.007234528,\n        -0.011330219,\n        -0.011446472,\n        0.02587976,\n        -0.017393272,\n        0.0055980403,\n        0.028133282,\n        -0.02122963,\n        0.012707373,\n        0.024949733,\n        -0.008119841,\n        -0.00077353115,\n        -0.021372711,\n        -0.0069528376,\n        -0.02850887,\n        0.0073015974,\n        0.01602506,\n        0.00612118,\n        -0.039740723,\n        0.00062374334,\n        0.01766155,\n        0.014773103,\n        -0.005043602,\n        0.037344117,\n        -0.024574146,\n        -0.015479566,\n        0.02584399,\n        0.019405348,\n        0.012367556,\n        -0.020228064,\n        0.008048301,\n        0.0033668727,\n        -0.03281918,\n        0.019065531,\n        -0.039883804,\n        -0.030279497,\n        -0.0032506194,\n        0.0046993135,\n        -0.007860507,\n        -0.031442028,\n        0.04424777,\n        0.012456981,\n        -0.018761484,\n        0.03913263,\n        -0.004386324,\n        0.011205023,\n        -0.001895822,\n        0.018475322,\n        -0.03571657,\n        -0.00074446783,\n        -0.010114032,\n        -0.032622445,\n        0.02360835,\n        0.003505482,\n        0.019244382,\n        0.028669836,\n        0.035895422,\n        0.03805952,\n        0.013601629,\n        -0.046966307,\n        -0.0053968327,\n        -0.029045423,\n        -0.033534586,\n        -0.012546407,\n        -0.016847776,\n        -0.0045316406,\n        -0.0122334175,\n        0.006134594,\n        -0.031638764,\n        0.00007650078,\n        0.01791194,\n        -0.017616836,\n        -0.016293338,\n        0.0071629877,\n        -0.018090792,\n        -0.010945689,\n        0.014889357,\n        0.026398428,\n        -0.024878193,\n        -0.0012698431,\n        0.019405348,\n        0.03414268,\n        0.031352602,\n        0.030494118,\n        -0.013404893,\n        -0.008964913,\n        -0.0005619838,\n        0.017241249,\n        0.00832552,\n        0.004627773,\n        0.008218209,\n        -0.014066641,\n        0.049863696,\n        -0.015747841,\n        0.034232106,\n        0.008575912,\n        0.018600518,\n        -0.02605861,\n        -0.023876626,\n        -0.024019707,\n        0.011446472,\n        0.044140458,\n        0.030297382,\n        0.035662916,\n        -0.0032998035,\n        -0.020031327,\n        0.012018796,\n        -0.028902343,\n        0.037344117,\n        0.010516447,\n        -0.018493207,\n        -0.029045423,\n        0.0013972744,\n        -0.014388573,\n        -0.017214421,\n        0.02668459,\n        0.024109133,\n        0.009389685,\n        0.007958875,\n        0.0031679007,\n        0.025486287,\n        0.0010323063,\n        0.07855142,\n        0.043389283,\n        0.01769732,\n        -0.0052403384,\n        -0.008960442,\n        0.0026537038,\n        0.015470623,\n        0.02112232,\n        0.0040979264,\n        -0.00061815424,\n        -0.07397283,\n        -0.022696208,\n        -0.011321276,\n        0.014907242,\n        -0.004596474,\n        0.007994645,\n        -0.0075922306,\n        0.004553997,\n        -0.004770854,\n        -0.03308746,\n        0.020478455,\n        -0.004788739,\n        -0.034393072,\n        -0.005638282,\n        0.016838834,\n        0.003686569,\n        -0.032801297,\n        0.029188504,\n        0.021927148,\n        0.002479324,\n        -0.0005589098,\n        -0.0061748354,\n        0.022517357,\n        0.012099279,\n        -0.008356819,\n        0.0062061343,\n        -0.005092786,\n        -0.001752741,\n        0.0064207558,\n        0.037522968,\n        -0.0033914647,\n        -0.0037133966,\n        -0.020979239,\n        -0.01588198,\n        0.00947911,\n        -0.0120098535,\n        0.035949077,\n        -0.018743599,\n        -0.002948808,\n        -0.017929826,\n        0.017223364,\n        -0.014120297,\n        -0.009362857,\n        0.016624212,\n        0.004473514,\n        -0.0042588925,\n        -0.015649473,\n        -0.0073194825,\n        -0.036807563,\n        0.005414718,\n        -0.050936803,\n        -0.039061088,\n        0.013208156,\n        0.0036038503,\n        0.018126562,\n        0.008406003,\n        -0.012886224,\n        0.006519124,\n        0.008803947,\n        0.026040725,\n        0.034643464,\n        0.001920414,\n        0.038703386,\n        -0.010292883,\n        0.009371799,\n        0.020907698,\n        -0.004609888,\n        -0.04249503,\n        0.010605873,\n        -0.0042521856,\n        0.00060250476,\n        -0.012126107,\n        0.0006148008,\n        0.012126107,\n        0.000029674617,\n        0.025790334,\n        -0.007489391,\n        0.009076695,\n        0.0017728618,\n        0.014737333,\n        -0.018904565,\n        -0.029796598,\n        0.009622191,\n        -0.000014854774,\n        0.0031366018,\n        -0.024466835,\n        0.005150913,\n        0.0051285564,\n        0.007954404,\n        0.013869905,\n        0.016096601,\n        -0.02584399,\n        0.008566969,\n        0.046250902,\n        0.010078262,\n        0.020371145,\n        -0.036771793,\n        -0.012644775,\n        -0.0006142419,\n        -0.03190704,\n        -0.023036025,\n        0.05483576,\n        0.006613021,\n        -0.012501694,\n        -0.001549298,\n        -0.000045131965,\n        -0.000108428496,\n        -0.021372711,\n        -0.023125451,\n        -0.03154934,\n        -0.01961997,\n        -0.016257567,\n        0.024055477,\n        0.026416313,\n        0.026165921,\n        -0.00306059,\n        -0.009917296,\n        -0.022624668,\n        0.012582177,\n        -0.0016387235,\n        0.024055477,\n        0.027095947,\n        0.00021601863,\n        -0.0017907469,\n        0.00323497,\n        -0.019977672,\n        -0.003822943,\n        -0.019298038,\n        0.019924017,\n        0.006215077,\n        0.03319477,\n        0.01815339,\n        -0.008003588,\n        0.019047646,\n        0.033659782,\n        -0.011839945,\n        -0.019298038,\n        0.0076905987,\n        0.0003873245,\n        0.016105544,\n        0.017679434,\n        -0.0048736935,\n        0.014173952,\n        0.030869706,\n        0.0097563295,\n        -0.010963574,\n        0.04303158,\n        -0.028025972,\n        0.010319711,\n        -0.00832552,\n        -0.025933415,\n        -0.021694642,\n        -0.03011853,\n        0.017402215,\n        -0.011831002,\n        -0.011106655,\n        -0.044033147,\n        0.0015750078,\n        -0.018976105,\n        -0.00493182,\n        0.0032506194,\n        0.009997779,\n        0.009344972,\n        -0.0064341696,\n        0.0027051233,\n        0.01752741,\n        -0.016150257,\n        -0.013270754,\n        -0.024931848,\n        -0.015381197,\n        -0.009675846,\n        0.025700908,\n        0.03396383,\n        -0.019476889,\n        0.022517357,\n        -0.027095947,\n        0.0109546315,\n        0.013279697,\n        -0.009657961,\n        0.00005554166,\n        -0.004976533,\n        0.009461225,\n        0.034446727,\n        0.021640986,\n        0.0023898983,\n        -0.012465924,\n        0.016463246,\n        0.02584399,\n        0.025021274,\n        0.015604761,\n        -0.02434164,\n        -0.009523823,\n        0.020084983,\n        0.0010563395,\n        -0.036914874,\n        0.00472167,\n        -0.015908808,\n        0.00433714,\n        -0.0041292254,\n        0.05537231,\n        0.02455626,\n        -0.020854043,\n        -0.016829891,\n        0.00017843195,\n        -0.034750775,\n        0.004690371,\n        -0.0076592998,\n        -0.028634066,\n        -0.0018902329,\n        0.020335374,\n        -0.029599862,\n        0.0077308402,\n        0.020907698,\n        0.039061088,\n        -0.0056025116,\n        0.0077800243,\n        -0.027471535,\n        -0.035144247,\n        0.012716316,\n        0.014084526,\n        0.004106869,\n        0.012448039,\n        0.020013442,\n        0.016570557,\n        -0.005687466,\n        0.0030024634,\n        0.017607894,\n        -0.01972728,\n        0.039776493,\n        0.0022389926,\n        -0.031048557,\n        -0.011008287,\n        0.0064967675,\n        -0.005848432,\n        -0.0073820804,\n        0.005767949,\n        0.0069528376,\n        0.022070229,\n        0.019280152,\n        0.0067874007,\n        0.010015664,\n        0.01675835,\n        -0.0049228775,\n        -0.0022747628,\n        0.020925583,\n        0.0174201,\n        0.028759262,\n        0.014442229,\n        0.009099051,\n        -0.00913035,\n        -0.034303647,\n        -0.017366445,\n        0.02536109,\n        0.019637855,\n        -0.046823226,\n        -0.0069081252,\n        -0.049470223,\n        -0.024270099,\n        -0.06817805,\n        -0.011303391,\n        0.011571668,\n        0.011240793,\n        0.019262267,\n        0.014433286,\n        0.01972728,\n        -0.011759462,\n        0.018001366,\n        0.004963119,\n        -0.0010848439,\n        -0.005168798,\n        -0.023107566,\n        0.008687694,\n        -0.019226497,\n        -0.02734634,\n        0.002367542,\n        0.009711617,\n        0.021658871,\n        0.01584621,\n        -0.028419444,\n        0.0011289978,\n        -0.001610778,\n        0.000898168,\n        0.026774015,\n        0.0126179475,\n        0.016293338,\n        -0.011598496,\n        0.03235417,\n        -0.029421011,\n        0.02788289,\n        0.015434853,\n        0.041851167,\n        -0.0082539795,\n        -0.005495201,\n        -0.0073954943,\n        -0.009657961,\n        -0.043246202,\n        0.006152479,\n        -0.013145559,\n        -0.044068918,\n        -0.025808219,\n        0.020854043,\n        0.019530544,\n        -0.019226497,\n        -0.032336283,\n        0.037487198,\n        0.015756784,\n        -0.020925583,\n        -0.013628457,\n        0.00042281527,\n        0.021712527,\n        -0.017777802,\n        -0.019119186,\n        -0.0102839405,\n        -0.006300031,\n        -0.003518896,\n        0.008651923,\n        -0.00003171813,\n        0.032443594,\n        -0.020514226,\n        0.0095327655,\n        -0.012868339,\n        0.030512003,\n        0.00056338107,\n        0.02452049,\n        0.0114733,\n        0.04070652,\n        -0.010650585,\n        0.010274998,\n        -0.017402215,\n        -0.010462792,\n        0.0026447612,\n        -0.03923994,\n        0.007534104,\n        -0.030440463,\n        0.010579045,\n        0.011401759,\n        -0.03414268,\n        0.0074178507,\n        -0.008271865,\n        0.020818273,\n        0.004726141,\n        -0.0010786959,\n        -0.034536153,\n        -0.011956198,\n        -0.020764617,\n        0.0019002933,\n        0.010122974,\n        -0.0033445163,\n        0.0065101814,\n        -0.0056203967,\n        -0.0045517613,\n        -0.010274998,\n        0.002693945,\n        -0.0056964085,\n        -0.0127610285,\n        0.052653775,\n        0.0021629808,\n        -0.008768177,\n        0.05515769,\n        0.004142639,\n        0.00013902881,\n        -0.01297565,\n        -0.009416512,\n        0.010650585,\n        -0.04435508,\n        -0.043210432,\n        0.01969151,\n        -0.014567425,\n        0.027239028,\n        -0.024395294,\n        0.008392589,\n        -0.0032126135,\n        -0.005311879,\n        -0.010435964,\n        0.019745165,\n        -0.017581066,\n        -0.0048602796,\n        0.0048110955,\n        0.041565005,\n        0.03008276,\n        0.0007522926,\n        -0.000059244438,\n        0.020210179,\n        0.016293338,\n        -0.01731279,\n        -0.020013442,\n        -0.016346993,\n        -0.015345427,\n        -0.00012582456,\n        -0.0076414146,\n        0.0054102466,\n        0.018457437,\n        -0.0081421975,\n        0.026165921,\n        0.009353914,\n        0.000045097033,\n        0.017133938,\n        0.027865006,\n        0.013091903,\n        0.016946144,\n        -0.0023094153,\n        0.013869905,\n        0.0045562326,\n        -0.05340495,\n        0.0045875316,\n        0.032300513,\n        -0.043711215,\n        -0.026344772,\n        -0.010677413,\n        -0.009496995,\n        0.028455215,\n        -0.00020525962,\n        0.012787856,\n        -0.030190071,\n        0.016901432,\n        0.026362658,\n        0.008455187,\n        -0.026148036,\n        -0.032443594,\n        -0.003599379,\n        0.014281263,\n        0.0044623357,\n        -0.00752069,\n        0.0047127274,\n        0.0326761,\n        0.020836158,\n        0.01815339,\n        0.0037894084,\n        0.0044891634,\n        -0.024395294,\n        0.018689943,\n        0.02360835,\n        0.0038900122,\n        0.011714749,\n        -0.013226042,\n        0.00668009,\n        0.0045383475,\n        0.027113833,\n        -0.031084327,\n        0.013243927,\n        0.019280152,\n        -0.036914874,\n        -0.023393728,\n        0.013860962,\n        -0.027024407,\n        0.011741577,\n        -0.031173753,\n        -0.010301826,\n        -0.0014296912,\n        -0.002174159,\n        0.022767749,\n        0.02049634,\n        -0.015184461,\n        0.01301142,\n        -0.011965141,\n        0.0025687495,\n        0.05841278,\n        0.002521801,\n        0.010274998,\n        0.038274143,\n        0.0004826186,\n        0.018457437,\n        0.036217354,\n        0.006076467,\n        0.0020858513,\n        0.0056338105,\n        0.015560049,\n        0.004815567,\n        0.020192293,\n        -0.037308346,\n        0.010695298,\n        0.020549996,\n        -0.013798365,\n        0.020621536,\n        -0.0073954943,\n        0.0128415115,\n        -0.00951488,\n        0.007757668,\n        -0.011696864,\n        0.006443112,\n        -0.01588198,\n        -0.036628712,\n        0.012609005,\n        -0.0009870346,\n        0.008942557,\n        0.042924274,\n        -0.033498816,\n        0.0035770226,\n        -0.021480022,\n        0.0010121856,\n        0.0005413041,\n        0.01776886,\n        0.042673882,\n        0.0031433087,\n        -0.016391706,\n        -0.04303158,\n        -0.057518523,\n        0.017956654,\n        0.03128106,\n        0.0022579955,\n        -0.01255535,\n        0.022499472,\n        0.0061793067,\n        0.020084983,\n        0.008656395,\n        0.0043326686,\n        -0.012197647,\n        0.010516447,\n        0.00052872865,\n        0.02521801,\n        0.027167488,\n        0.0008534553,\n        -0.04235195,\n        -0.008245037,\n        0.013637399,\n        -0.01339595,\n        -0.027292684,\n        -0.0072881836,\n        0.026398428,\n        -0.0016074246,\n        0.030100645,\n        -0.021319056,\n        -0.043746985,\n        0.014361746,\n        -0.0237872,\n        -0.006943895,\n        -0.030619314,\n        0.068893455,\n        -0.0014073348,\n        -0.005884202,\n        0.040026885,\n        -0.034017485,\n        -0.05487153,\n        0.011920428,\n        0.0063671004,\n        -0.0017337381,\n        0.027256913,\n        -0.0021719234,\n        0.013467491,\n        0.03941879,\n        0.032121662,\n        0.019387463,\n        0.030207956,\n        -0.023429498,\n        0.0046814284,\n        -0.022231195,\n        -0.013458548,\n        0.016525844,\n        -0.012993535,\n        -0.009166121,\n        0.013002478,\n        -0.0028593824,\n        -0.0009372917,\n        0.022356391,\n        0.0025933415,\n        -0.0070959185,\n        0.014764161,\n        -0.044998944,\n        0.01496984,\n        0.0010295118,\n        0.021909263,\n        0.010480677,\n        -0.00030320857,\n        -0.016293338,\n        -0.029224275,\n        0.013038248,\n        0.010516447,\n        -0.030547773,\n        -0.041565005,\n        0.028312134,\n        0.008307635,\n        0.017160766,\n        0.004173938,\n        -0.018278586,\n        -0.012152934,\n        0.0054102466,\n        0.0059199724,\n        0.009336029,\n        -0.04313889,\n        0.003957081,\n        -0.018940335,\n        -0.014254435,\n        -0.021837723,\n        0.0021752769,\n        0.012376498,\n        -0.0026693533,\n        -0.011911485,\n        0.013547974,\n        0.027203258,\n        0.004019679,\n        -0.003648563,\n        0.010534332,\n        -0.029027538,\n        0.0041269897,\n        -0.0026067554,\n        0.020371145,\n        0.012743143,\n        0.035340983,\n        0.05476422,\n        0.026219577,\n        -0.022392161,\n        -0.01811762,\n        0.009657961,\n        0.029635632,\n        -0.01692826,\n        -0.019852476,\n        -0.02115809,\n        -0.0007657064,\n        0.03128106,\n        -0.016427476,\n        -0.02452049,\n        0.0237872,\n        -0.031513568,\n        0.026255347,\n        -0.028061742,\n        0.02375143,\n        -0.016955087,\n        -0.019512659,\n        -0.010650585,\n        -0.008996212,\n        0.015721014,\n        0.0020400207,\n        -0.0105879875,\n        -0.037272576,\n        0.00015761257,\n        -0.014844644,\n        -0.05029294,\n        0.0052939937,\n        -0.007408908,\n        0.00018751423,\n        0.024663571,\n        -0.038703386,\n        0.027471535,\n        -0.005463902,\n        -0.009014097,\n        -0.004455629,\n        0.0077308402,\n        0.020478455,\n        0.03920417,\n        -0.049935237,\n        0.0042790133,\n        -0.015980348,\n        -0.028383674,\n        0.0036686838,\n        0.005866317,\n        -0.010292883,\n        -0.005186683,\n        0.017563181,\n        -0.025021274,\n        -0.005530971,\n        -0.012752086,\n        -0.031227408,\n        -0.026523624,\n        -0.02664882,\n        0.0016689047,\n        -0.016069774,\n        -0.004305841,\n        -0.0036284423,\n        0.0000577773,\n        -0.012948822,\n        -0.0044578644,\n        -0.0058707884,\n        -0.038703386,\n        -0.0064207558,\n        -0.025522057,\n        0.010427021,\n        0.011339162,\n        -0.010292883,\n        0.018475322,\n        -0.0048736935,\n        0.015756784,\n        0.006376043,\n        0.01143753,\n        0.025343206,\n        0.0034249993,\n        -0.01378048,\n        0.020210179,\n        -0.0004591444,\n        -0.0045383475,\n        0.012778914,\n        -0.019065531,\n        -0.004037564,\n        0.038846467,\n        0.010981459,\n        -0.00913035,\n        0.008611682,\n        0.010963574,\n        -0.017885113,\n        0.028634066,\n        0.029707173,\n        0.012456981,\n        -0.030547773,\n        0.002503916,\n        0.021247515,\n        0.0047753253,\n        -0.010713183,\n        -0.0077710818,\n        0.04102845,\n        0.009953066,\n        0.012081394,\n        -0.0088620735,\n        -0.027078062,\n        0.011249736,\n        -0.02448472,\n        0.00041974126,\n        -0.03070874,\n        0.022445817,\n        -0.0005066517,\n        0.026845556,\n        0.0015235881,\n        -0.01976305,\n        0.0075877593,\n        0.027918661,\n        -0.01504138,\n        0.009747387,\n        0.003925782,\n        0.011768404,\n        -0.015756784,\n        0.019745165,\n        0.014808874,\n        0.016624212,\n        0.03913263,\n        0.017974539,\n        0.017017685,\n        -0.016141314,\n        -0.0009473521,\n        0.013574801,\n        0.000012182487,\n        -0.014004043,\n        -0.009881525,\n        0.004565175,\n        0.038882237,\n        -0.004900521,\n        -0.046823226,\n        0.0054102466,\n        -0.040420357,\n        -0.0039235465,\n        -0.0076995413,\n        -0.015747841,\n        -0.01633805,\n        -0.028920228,\n        0.008124312,\n        0.013896733,\n        0.003116481,\n        -0.028097512,\n        -0.027990201,\n        0.019852476,\n        0.005204568,\n        0.038846467,\n        0.021640986,\n        0.009649019,\n        -0.00014685355,\n        0.0028079627,\n        -0.027990201,\n        -0.0041158115,\n        0.004375146,\n        0.01465685,\n        -0.017330674,\n        -0.043389283,\n        0.015023495,\n        -0.033212654,\n        0.013279697,\n        0.013860962,\n        0.028723491,\n        0.004963119,\n        0.022749864,\n        0.016132372,\n        -0.0037514025,\n        0.001164768,\n        -0.008419417,\n        0.011526955,\n        -0.002392134,\n        0.005499672,\n        -0.017554238,\n        0.02532532,\n        0.0037022184,\n        -0.017500583,\n        -0.0017918647,\n        -0.010158745,\n        -0.004176174,\n        0.01504138,\n        -0.006089881,\n        -0.00027163015,\n        -0.03308746,\n        0.01504138,\n        0.0014218665,\n        -0.011401759,\n        0.010087204,\n        0.03305169,\n        0.0015481801,\n        0.03956187,\n        0.02609438,\n        -0.010829436,\n        0.028938113,\n        0.00004279153,\n        -0.00058014836,\n        -0.014218665,\n        0.0030963602,\n        0.043568134,\n        0.003364637,\n        0.008987269,\n        -0.03432153,\n        0.00853567,\n        0.017983481,\n        -0.014916184,\n        -0.005119614,\n        0.011339162,\n        0.0036016146,\n        0.0038698914,\n        -0.00710039,\n        -0.005168798,\n        -0.016436419,\n        -0.03081605,\n        0.0078023807,\n        0.0133691225,\n        0.012895167,\n        0.023733545,\n        0.034446727,\n        0.018528977,\n        0.01811762,\n        0.012287073,\n        0.03480443,\n        -0.0034786544,\n        0.007489391,\n        -0.013521146,\n        0.0135926865,\n        -0.008656395,\n        -0.03242571,\n        0.0033266312,\n        0.007547518,\n        -0.00084674836,\n        -0.0044891634,\n        -0.01640959,\n        -0.013628457,\n        0.059736278,\n        0.011759462,\n        0.024073362,\n        -0.0008814007,\n        -0.046644375,\n        0.01591775,\n        -0.02232062,\n        -0.02375143,\n        0.01899399,\n        0.003456298,\n        -0.0013022599,\n        -0.02232062,\n        -0.0033691083,\n        -0.005884202,\n        -0.005495201,\n        -0.0030807108,\n        0.02228485,\n        0.0073507815,\n        0.021408482,\n        0.015631588,\n        -0.029027538,\n        0.017223364,\n        -0.009282374,\n        0.010704241,\n        0.0050838436,\n        0.013664227,\n        0.0032617976,\n        -0.019083416,\n        -0.027578846,\n        0.026505738,\n        -0.0130650755,\n        0.023876626,\n        0.013494318,\n        -0.02993968,\n        0.03777336,\n        0.0035814939,\n        0.045535497,\n        0.0035792582,\n        0.013869905,\n        -0.013109788,\n        -0.011750519,\n        0.02672036,\n        0.031191638,\n        -0.017750975,\n        -0.023465268,\n        0.042745423,\n        0.0084775435,\n        0.008651923,\n        0.04575012,\n        0.029635632,\n        0.038703386,\n        0.00080203556,\n        -0.030941246,\n        -0.024717227,\n        0.034393072,\n        -0.0040554493,\n        -0.0071719303,\n        0.009970951,\n        0.010194515,\n        -0.0014486942,\n        -0.04317466,\n        -0.03501905,\n        0.019709395,\n        0.0031254236,\n        -0.012394384,\n        -0.019870361,\n        -0.014943012,\n        0.005347649,\n        -0.02459203,\n        0.010614815,\n        0.026326887,\n        0.03469712,\n        -0.012671603,\n        -0.014495884,\n        -0.020621536,\n        0.019959787,\n        -0.0015716543,\n        0.0038631845,\n        -0.0041828807,\n        0.040277276,\n        0.0071585164,\n        -0.021980803,\n        -0.015729956,\n        -0.00055779197,\n        0.025647253,\n        0.036736023,\n        0.0139593305,\n        -0.0059557427,\n        -0.006170364,\n        -0.03070874,\n        -0.040134195,\n        0.025074929,\n        -0.008987269,\n        -0.008736878,\n        -0.0030762395,\n        0.006022812,\n        -0.024627801,\n        -0.018976105,\n        0.012358613,\n        0.039883804,\n        0.0012910817,\n        -0.008562498,\n        0.010856263,\n        0.024252214,\n        -0.015926693,\n        0.015479566,\n        -0.022499472,\n        -0.064243324,\n        0.0100603765,\n        0.0061256513,\n        -0.027435765,\n        0.009237661,\n        0.04310312,\n        0.032336283,\n        0.0024860308,\n        0.0018444023,\n        -0.010373366,\n        -0.013860962,\n        0.043890066,\n        -0.011240793,\n        0.030529888,\n        0.017536353,\n        -0.005897616,\n        -0.026559394,\n        -0.016633155\n      ]\n    }\n  ],\n  \"model\": \"text-search-babbage:001\",\n  \"usage\": {\n    \"prompt_tokens\": 8,\n    \"total_tokens\": 8\n  }\n}\n".data(using: .utf8)
+
         let result = try await openAI.embeddings(query: query)
+
+        XCTAssertEqual(mockURLSession.lastRequest?.allHTTPHeaderFields, ["Content-Type": "application/json", "Authorization": "Bearer <MOCK_API_TOKEN>"])
         XCTAssertFalse(result.data.isEmpty)
+    }
+
+    func testChatsAsync() async throws {
+        let query = OpenAI.ChatQuery(model: .gpt4, messages: [
+            .init(role: .system, content: "You are Librarian-GPT. You know everything about the books."),
+            .init(role: .user, content: "Who wrote Harry Potter?")
+        ])
+
+        mockURLSession.mockData = "{\"id\":\"chatcmpl-6zyw8Grs9kzUzPFaPmoquaw0n1AJF\",\"object\":\"chat.completion\",\"created\":1680230268,\"model\":\"gpt-4-0314\",\"usage\":{\"prompt_tokens\":30,\"completion_tokens\":72,\"total_tokens\":102},\"choices\":[{\"message\":{\"role\":\"assistant\",\"content\":\"The Harry Potter series was written by British author J.K. Rowling, whose full name is Joanne Rowling. The series consists of seven books, starting with \\\"Harry Potter and the Philosopher\'s Stone\\\" (published as \\\"Harry Potter and the Sorcerer\'s Stone\\\" in the United States) and ending with \\\"Harry Potter and the Deathly Hallows.\\\"\"},\"finish_reason\":\"stop\",\"index\":0}]}\n".data(using: .utf8)
+
+        let result = try await openAI.chats(query: query)
+
+        XCTAssertFalse(result.choices.isEmpty)
     }
 
     func testSimilarity_Similar() {
@@ -49,17 +90,8 @@ final class OpenAITests: XCTestCase {
 
     func testMakeRequest_bodyContainsFractional() throws {
         let query = OpenAI.CompletionsQuery(model: .textDavinci_003, prompt: "What is 42?", temperature: 0.7, maxTokens: 100)
-        let request = try openAI.makeRequest(query: query, url: .completions, timeoutInterval: 60.0)
+        let request = try OpenAIImpl.makeRequest(query: query, url: .completions, timeoutInterval: 60.0, apiToken: "<API_TOKEN>")
         let json = try JSONSerialization.jsonObject(with: request.httpBody!, options: []) as! [String: Any]
         XCTAssertEqual(json["temperature"] as? Double, 0.7)
-    }
-    
-    func testChatsAsync() async throws {
-        let query = OpenAI.ChatQuery(model: .gpt4, messages: [
-            .init(role: .system, content: "You are Librarian-GPT. You know everything about the books."),
-            .init(role: .user, content: "Who wrote Harry Potter?")
-        ])
-        let result = try await openAI.chats(query: query)
-        XCTAssertFalse(result.choices.isEmpty)
     }
 }
